backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.3
beautifulsoup4==4.13.4
readability-lxml==0.8.4.1
python-dotenv==1.1.1
openai==1.99.6
